backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.6
dataclasses-json==0.6.7
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.12
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-ollama==1.0.1
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
marshmallow==3.26.2
multidict==6.7.1
mypy-extensions==1.1.0
numpy==2.4.3
ollama==0.6.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pgvector==0.4.2
pip==26.0.1
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
pyyaml==6.0.3
report==0.0.1
requests==2.32.5
requests-toolbelt==1.0.0
sqlalchemy==2.0.48
tenacity==9.1.4
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0